requirements.txt
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0
scikit-learn>=1.3.0
google-genai>=0.3.0
google-generativeai>=0.3.0
opencv-python-headless>=4.8.0
pyinstaller>=5.13.0
grpcio>=1.50.0
protobuf>=4.21.0
google-auth>=2.15.0
google-auth-oauthlib>=0.7.0
google-auth-httplib2>=0.1.0